pyproject.toml
[tool.poetry]
name = "family_doctor"
version = "0.1.0"
description = "Test project"
authors = ["Bazarovinc"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.73.0"
pydantic = {extras = ["dotenv"], version = "^1.9.0"}
uvicorn = "^0.17.4"
dependency-injector = "^4.38.0"
tortoise-orm = "^0.18.1"
asyncpg = "^0.25.0"
aerich = "^0.6.2"
requests = "^2.27.1"
geopy = "^2.2.0"


[tool.poetry.dev-dependencies]


[tool.aerich]
tortoise_orm = "core.database.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
